+++ src/task/taskModule.ts
@@ -4,7 +4,7 @@ import type {ModuleMeta, LoadModuleResult} from '../fs/modules.js';
 import {toTaskName, isTaskPath, TASK_FILE_SUFFIX} from './task.js';
 import type {Task} from './task.js';
 import {getPossibleSourceIds} from '../fs/inputPath.js';
-import type {Obj} from '../index.js';
+import type {Obj} from '../utils/types.js';
 import type {Filesystem} from '../fs/filesystem.js';
 
 export interface TaskModule {